Move private static classes(or relevant functions) out of DoubleValuesSource
